check point 2 v2.0 start
Former-commit-id: 6b015121f03240cd95b64a3c5cf57de8e011e595

+++ src/dao/Datum.java
@@ -26,7 +26,7 @@ public abstract class Datum implements Comparable<Datum> {
 			case Bool:
 				DatumBool valBool = (DatumBool)this;
 				return String.valueOf(valBool.getValue());
-			case Float:
+			case Double:
 				DatumDouble valFloat = (DatumDouble)this;
 				return valFloat.toString();
 			default:	//Int or Long
